remove leaderboard function from scores
