+++ src/index.js
@@ -1,6 +1,17 @@
 import './style.css';
+import scores from './scores';
 
-import leaderboard from './scores';
+const leaderboard = () => {
+  let newScore = '';
+  scores.forEach((leader) => {
+    newScore += `
+    <li>
+      <span>${leader.name}</span>
+      <span>${leader.score}</span>
+    </li>`;
+  });
+
+  document.querySelector('.board').innerHTML = newScore;
+}
 
 leaderboard();
-
